requirements.txt
discord.py
time